Conflicts:
	test/Histria.Model.Tests/DerivedClasses.cs

+++ test/Histria.Model.Tests/DerivedClasses.cs
@@ -18,19 +18,19 @@ namespace Histria.Model.Tests.ModelToTest
         public virtual string ParentMember { get; set; }
     }
 
-    [Inheritance("TypeClass", (int)EnumClassType.CC1)]
+    [Inheritance("TypeClass", EnumClassType.CC1)]
     public class CC1 : CParent
     {
         public virtual string CC1Member { get; set; }
     }
 
-    [Inheritance("TypeClass", (int)EnumClassType.CC2)]
+    [Inheritance("TypeClass", EnumClassType.CC2)]
     public class CC2 : CParent
     {
         public virtual string CC2Member { get; set; }
     }
 
-    [Inheritance("TypeClass", (int)EnumClassType.CC3)]
+    [Inheritance("TypeClass", EnumClassType.CC3)]
     public class CC3 : CC1
     {
         public virtual string CC3Member { get; set; }
@@ -48,10 +48,11 @@ namespace Histria.Model.Tests.ModelToTest
         public virtual EnumClassType TypeClass { get; set; }
     }
     [Index("I1,I3")]
-    [Inheritance("TypeClass", (int)EnumClassType.CC4)]
+    [Inheritance("TypeClass", EnumClassType.CC4)]
     public class CC4 : CP1
     {
         public virtual string I3 { get; set; }
+
     }
 
 
@@ -65,7 +66,7 @@ namespace Histria.Model.Tests.ModelToTest
     }
 
     [Db("YYY")]
-    [Inheritance("TypeClass", (int)EnumClassType.CC5)]
+    [Inheritance("TypeClass", EnumClassType.CC5)]
     public class CC5 : CP2
     {
     }
@@ -82,7 +83,7 @@ namespace Histria.Model.Tests.ModelToTest
     }
 
     [PrimaryKey("Name")]
-    [Inheritance("TypeClass", (int)EnumClassType.CC6)]
+    [Inheritance("TypeClass", EnumClassType.CC6)]
     public class CC6 : CP3
     {
         public virtual string Name { get; set; }
@@ -100,7 +101,7 @@ namespace Histria.Model.Tests.ModelToTest
         public virtual string Id { get; set; }
     }
 
-    [Inheritance("TypeClass", (int)EnumClassType.CC7)]
+    [Inheritance("TypeClass", EnumClassType.CC7)]
     public class CC7 : CP4
     {
         public virtual string Name { get; set; }
